Added extra warnings in CMake build (#31)
* Added extra warnings in CMake build

* Fix errors generated with strict settings on MSVC

* Fix errors generated with strict settings for gcc

---------

+++ epoch_simulator/epoch_lib/Optimisation/LeagueTable.hpp
@@ -31,7 +31,7 @@ public:
 
 
 private:
-	int mCapacity;
+	size_t mCapacity;
 	OptimiserConfig mConfig;
 
 	void considerMinimum(std::multimap<float, uint64_t>& subTable, float value, uint64_t paramIndex);
